refactor(reputation): migrate reputation into core addon
