+++ addons/core/functions/fn_addVehicleKeyWest.sqf
@@ -3,6 +3,6 @@
 params ["_target", "_caller"];
 
 _caller addItem "ACE_key_west";
-_target commandChat localize "str.coopr.init.chatter.keys";
+_target commandChat localize "str.coopr.core.chatter.keys";
 
 FLOG("%1 has requested a vehicle key", name _caller);
