+++ addons/core/functions/fn_convertTempToReputation.sqf
