+++ addons/core/init/fn_initCommanderNPC.sqf
