+++ addons/core/init/fn_initCoopRClients.sqf
