+++ addons/core/init/fn_initCoopRServer.sqf
