+++ addons/core/init/fn_initEventsClient.sqf
