+++ addons/core/init/fn_initEventsServer.sqf
