+++ addons/core/init/fn_initHQ.sqf
