+++ addons/core/init/fn_initIntelligenceNPC.sqf
