+++ addons/core/init/fn_initQuartermasterNPC.sqf
