+++ addons/core/init/fn_setupBasicRoleLoadoutModule.sqf
