+++ addons/core/init/fn_setupCommanderModule.sqf
