+++ addons/core/init/fn_setupHQModule.sqf
