+++ addons/core/init/fn_setupIntelligenceModule.sqf
