+++ addons/core/init/fn_setupQuartermasterModule.sqf
