+++ addons/core/init/fn_setupVehicleModule.sqf
