+++ addons/core/init/script_component.hpp
