+++ addons/core/script_component.hpp
@@ -5,5 +5,6 @@
 #define DEBUG_CTX "COOPR.CORE"
 
 #define SYNC_RATE_MIN 3
+#define COOPR_REP_AAR_MULTIPLIER 1.0
 
 #include "\x\coopr\addons\core\script_macros.hpp"
