+++ addons/core/stringtable.xml
@@ -1,19 +1,19 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <Project name="CoopR Mod">
-    <Package name="Init">
-        <Key ID="str.coopr.init.action.vehicles">
+    <Package name="Core">
+        <Key ID="str.coopr.core.action.vehicles">
             <Original>Get vehicle keys</Original>
             <German>Fahrzeugschluessel anfordern</German>
         </Key>
-        <Key ID="str.coopr.init.chatter.keys">
+        <Key ID="str.coopr.core.chatter.keys">
             <Original>Please drive carefully. The technical maintenance is due!</Original>
             <German>Bitte fahrt vorsichtig. Die TMP ist bald wieder dran!</German>
         </Key>
-        <Key ID="str.coopr.init.action.rep">
+        <Key ID="str.coopr.core.action.rep">
             <Original>Show reputation</Original>
             <German>Ruf anzeigen</German>
         </Key>
-        <Key ID="str.coopr.init.action.intel">
+        <Key ID="str.coopr.core.action.intel">
             <Original>Show intel status</Original>
             <German>Aufklaerungsstand anzeigen</German>
         </Key>
@@ -25,5 +25,13 @@
             <Original>Equipment</Original>
             <German>Ausruestung</German>
         </Key>
+            <Key ID="str.coopr.aar.action.commander">
+                <Original>After Action Report</Original>
+                <German>Abschlussbericht</German>
+            </Key>
+            <Key ID="str.coopr.reputation.chatter.aar">
+                <Original>Good Job! Glad you made it back!</Original>
+                <German>Gute Arbeit! Wir sind froh, dass ihr es zurueck geschafft habt!</German>
+            </Key>
     </Package>
 </Project>
