+++ addons/persistence/functions/fn_sync.sqf
