+++ addons/persistence/functions/fn_syncPlayerToServer.sqf
