--- addons/reputation/CfgFunctions.hpp
@@ -1,16 +1,0 @@
-class CfgFunctions {
-    class coopr {
-        class coopr_reputation_functions {
-            file = "x\coopr\addons\reputation\functions";
-            class updateReputation {};
-            class updateTempReputation {};
-            class convertTempToReputation {};
-            class deliverAfterActionReport {};
-            class showReputation {};
-        }
-        class coopr_reputation_tests {
-            file = "x\coopr\addons\reputation\tests";
-            class tests_reputation {};
-        }
-    }
-}
