--- addons/reputation/script_component.hpp
@@ -1,9 +1,0 @@
-#include "\x\coopr\addons\core\script_mod.hpp"
-
-#define COMPONENT reputation
-
-#define DEBUG_CTX "COOPR.REPUTATION"
-
-#define COOPR_REP_AAR_MULTIPLIER 1.0
-
-#include "\x\coopr\addons\core\script_macros.hpp"
