+++ build.sh
@@ -17,7 +17,6 @@ $ARMAKE build -p -f addons/eventhandler $DEST/eventhandler.pbo
 $ARMAKE build -p -f addons/integration $DEST/integration.pbo
 $ARMAKE build -p -f addons/intel $DEST/intel.pbo
 $ARMAKE build -p -f addons/lobby $DEST/lobby.pbo
-$ARMAKE build -p -f addons/reputation $DEST/reputation.pbo
 $ARMAKE build -p -f addons/supply $DEST/supply.pbo
 
 cp mod.cpp @CoopR/mod.cpp
